requirements.txt
python-telegram-bot>=11.1.0
configparser>=3.5.0

